pyproject.toml
[project]
name = "cinereserve-api"
version = "0.1.0"
description = "Backend API for Cinépolis Natal"
authors = [
    {name = "Ramon Jales ramonjales123@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.17.0,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "redis (>=7.3.0,<8.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "gunicorn (>=25.1.0,<26.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "environs[django] (>=14.6.0,<15.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.7,<0.16.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]

[tool.ruff]
line-length = 100
exclude = [".git", ".venv", "migrations", "__pycache__"]

[tool.ruff.lint]
# Pyflakes (F), pycodestyle (E), isort (I), pyupgrade (UP)
select = ["E", "F", "I", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"